+++ app/actions/keys.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs-extra';
 import * as sshpk from 'sshpk';
 import {toastr} from 'react-redux-toastr';
-import { remote } from 'electron';
+import { clipboard, remote } from 'electron';
 import {GetState, OptionalAction, TypedAction, TypedThunk} from './types/index';
 import {FormState, State} from './types/keys';
 import {afterAction} from '../store/eventMiddleware';
@@ -186,6 +186,14 @@ export function loadPrivateKey(): Thunk<void> {
     }
 
     dispatch(change({ ...keys.formState, publicKey: privateKey.key.toPublic().toString('ssh') }));
+  };
+}
+
+export function copyToClipboard(username: string): Thunk<void> {
+  return (dispatch, getState) => {
+    const { keys } = getState();
+    clipboard.writeText(keys.byUser[username].toString('ssh'));
+    toastr.success('', `Public key of ${username} copied to clipboard`, { timeOut: 2000 });
   };
 }